% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_compare_splicing.R
\name{calculate_unique_events}
\alias{calculate_unique_events}
\title{Identify Unique Splicing Events}
\usage{
calculate_unique_events(events_dt, family_pct_cols)
}
\arguments{
\item{events_dt}{A data.table with splicing event data}

\item{family_pct_cols}{Character vector of family pct column names}
}
\value{
Character vector indicating family uniqueness per event
}
\description{
Determines which splicing events are present in family members but absent in
controls (control average is 0 and family member value is non-zero).
}
\keyword{internal}
