% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_run_cortar.R
\name{cortar}
\alias{cortar}
\title{Run cortar}
\usage{
cortar(
  file,
  mode = "default",
  assembly = "hg38",
  annotation = "UCSC",
  input_type = "bamfile",
  paired = TRUE,
  stranded = 2,
  subset = NULL,
  output_dir = "~",
  genelist = NULL,
  prefix = "",
  debug = FALSE,
  ria = TRUE
)
}
\arguments{
\item{file}{A file path, pointing to the cortar samplefile - see readme for
more information}

\item{mode}{One of \code{"default"} (default),\code{"panel"}, or \code{"research"} - see
readme for more information}

\item{assembly}{Assembly used for alignment: either \code{"hg38"} (default) or
\code{"hg19"}}

\item{annotation}{Annotation used for alignment: either \code{"UCSC"} or
\code{"1000genomes"}}

\item{paired}{Is the RNA-seq paired-end?: \code{TRUE}/\code{FALSE}}

\item{stranded}{Strandedness of the RNA-seq: \code{0} for unstranded, \code{1}
for forward stranded or \code{2} for reverse stranded}

\item{subset}{Does the RNA-Seq need to be subsetted to the genes
of interest? \code{TRUE}/\code{FALSE} (Optional but significantly improves speed
of subsequent analyses - not currently available)}

\item{output_dir}{A directory path, pointing to the desired location for
export of cortar results (e.g. \code{"output/"})}

\item{genelist}{A character vector with genes/RefSeq transcripts of interest
(Only for panel or research modes; default = NULL)}

\item{prefix}{A character vector to be appended to the beginning of the
output file names}
}
\description{
\code{cortar()} runs the entire cortar pipeline with specified parameters,
and returns excel & pdf reports.
}
\examples{
#### == COMING SOON == ####

}
