% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_run_cortar.R
\name{cortar_batch}
\alias{cortar_batch}
\title{Run multiple instances of cortar}
\usage{
cortar_batch(
  folder,
  pattern = "*.tsv",
  mode = "default",
  assembly = "hg38",
  annotation = "UCSC",
  input_type = "sj",
  paired = TRUE,
  stranded = 2,
  subset = FALSE,
  output_dir = "~",
  genelist = NULL,
  prefix = "",
  debug = FALSE,
  ria = TRUE
)
}
\arguments{
\item{folder}{A character vector of the full file path, pointing to a folder
containing one or more cortar samplefiles}

\item{pattern}{An optional regular expression. Only file names which match
the regular expression will be returned (default: \code{"*.tsv"}).}

\item{mode}{One of \code{"default"} (default),\code{"panel"}, or \code{"research"} - see
readme for more information}

\item{assembly}{Assembly used for alignment: either \code{"hg38"} (default) or
\code{"hg19"}}

\item{annotation}{Annotation used for alignment: either \code{"UCSC"} or
\code{"1000genomes"}}

\item{paired}{Is the RNA-seq paired-end?: \code{TRUE}/\code{FALSE}}

\item{stranded}{Strandedness of the RNA-seq: \code{0} for unstranded, \code{1}
for forward stranded or \code{2} for reverse stranded}

\item{subset}{Does the RNA-Seq need to be subsetted to the genes
of interest? \code{TRUE}/\code{FALSE} (Optional but significantly improves speed
of subsequent analyses - not currently available)}

\item{output_dir}{A directory path, pointing to the desired location for
export of cortar results (e.g. \code{"output/"})}

\item{genelist}{A character vector with genes/RefSeq transcripts of interest
(Only for panel or research modes; default = NULL)}

\item{prefix}{A character vector to be appended to the beginning of the
output file names}
}
\description{
\code{cortar_batch()} runs the entire cortar pipeline with specified parameters,
and returns excel & pdf reports for multiple samplefiles.
}
\examples{
#### == COMING SOON == ####

}
