% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_extract_count_reads.R
\name{extractCountReads}
\alias{extractCountReads}
\title{Extract and count split-reads from .bam files}
\usage{
extractCountReads(
  genes.GRanges,
  introns.GRanges,
  intron_starts.GRanges,
  intron_ends.GRanges,
  bamfiles,
  sjfiles,
  irfiles,
  sample_names,
  assembly,
  annotation,
  paired,
  stranded,
  input,
  debug
)
}
\arguments{
\item{genes.GRanges}{A GRanges object containing the coordinates of the
genes of interest. Created by \code{selectGenesTranscripts()}}

\item{introns.GRanges}{A GRanges object containing the coordinates of the
introns of the genes of interest. Created by \code{selectGenesTranscripts()}}

\item{intron_starts.GRanges}{A GRanges object containing the coordinates
of the upstream exon-intron junctions of the genes of interest.
Created by \code{selectGenesTranscripts()}}

\item{intron_ends.GRanges}{A GRanges object containing the coordinates
of the downstream exon-intron junctions of the genes of interest.
Created by \code{selectGenesTranscripts()}}

\item{bamfiles}{A character vector of file paths, pointing to the .bam files
for analysis. Included in a properly formatted cortar samplefile.}

\item{sample_names}{A character vector of sample names. Included in a
properly formatted cortar samplefile.}

\item{assembly}{Assembly used for alignment: either \code{"hg38"} or \code{"hg19"}}

\item{annotation}{Annotation used for alignment: either \code{"UCSC"} or
\code{"1000genomes"}}

\item{paired}{Is the RNA-seq paired-end?: \code{TRUE}/\code{FALSE}}

\item{stranded}{Strandedness of the RNA-seq: \code{0} for unstranded, \code{1}
for forward stranded or \code{2} for reverse stranded}
}
\value{
A GRanges object consisting of aggregated junctional read count data
from all control and test samples. The GRanges object has at least three
metadata columns:
\itemize{
\item SJ_IR: whether the entry is counting split (SJ) or non-split (IR) reads at
the junction.
\item gene: the name of the gene which the junction occurs within or \code{NA} if the
junction is not with a gene
*count_sample_name: the count for each junction for a given sample. There
will be as many count_sample_name metadata columns as there are
sample_names specified
}
}
\description{
\code{selectGenesTranscripts} extracts split-reads from control and test .bam
files, counts them and aggregates the data into a single data table.
}
\examples{
#### == COMING SOON == ####

}
