% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_generate_samplefile.R
\name{generate_sample_metadata}
\alias{generate_sample_metadata}
\title{Generate Sample Metadata}
\usage{
generate_sample_metadata(
  sampleID,
  familyID,
  gene,
  transcript,
  alignment,
  bamfile_directory,
  sjir_directory,
  provider,
  RNAseq,
  assembly,
  age_of_biopsy,
  sex,
  tissue,
  chx_dmso
)
}
\arguments{
\item{sampleID}{Sample identifier}

\item{familyID}{Family identifier}

\item{gene}{Gene name}

\item{transcript}{Transcript identifier}

\item{alignment}{Alignment identifier}

\item{bamfile_directory}{Directory containing BAM files}

\item{sjir_directory}{Directory containing splice junction files}

\item{provider}{Data provider name}

\item{RNAseq}{RNA-seq type}

\item{assembly}{Genome assembly version}

\item{age_of_biopsy}{Age at biopsy}

\item{sex}{Sample sex}

\item{tissue}{Tissue type}

\item{chx_dmso}{Treatment condition}
}
\value{
A data.table with sample metadata
}
\description{
Creates a data.table with sample metadata for generating a cortar samplefile.
}
\keyword{internal}
