% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_debug_enabled}
\alias{is_debug_enabled}
\title{Check if debug mode is enabled}
\usage{
is_debug_enabled(debug)
}
\arguments{
\item{debug}{Debug parameter (can be a path string or FALSE)}
}
\value{
TRUE if debug mode is enabled, FALSE otherwise
}
\description{
Check if debug mode is enabled
}
\keyword{internal}
