% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_extract_count_reads.R
\name{read_bam_sample}
\alias{read_bam_sample}
\title{Read BAM File for a Single Sample}
\usage{
read_bam_sample(
  bamfile,
  param,
  paired,
  stranded,
  intron_starts.GRanges,
  intron_ends.GRanges,
  introns.GRanges,
  Genome_Assembly
)
}
\arguments{
\item{bamfile}{Path to a BAM file}

\item{param}{ScanBamParam object used for read extraction}

\item{paired}{Logical indicating paired-end sequencing}

\item{stranded}{Strand mode for paired-end reads}

\item{intron_starts.GRanges}{GRanges of intron start junctions}

\item{intron_ends.GRanges}{GRanges of intron end junctions}

\item{introns.GRanges}{GRanges of introns}

\item{Genome_Assembly}{BSgenome object for junction summarisation}
}
\value{
List with \code{sj} (splice junction GRanges) and \code{ir} (intron GRanges)
}
\description{
Internal helper that extracts split-read and intron-retention counts from a
single BAM file.
}
\keyword{internal}
