% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_run_cortar.R
\name{run_cortar_test}
\alias{run_cortar_test}
\title{Run the cortar test}
\usage{
run_cortar_test(test_path = getwd())
}
\arguments{
\item{test_path}{Character string indicating the path to the directory where the test will run.
By default, it uses the current working directory (\code{getwd()}).}
}
\value{
None. The function will perform the tests and modify files, but does not return any values.
}
\description{
This function runs a test for the \code{cortar} package by copying the necessary files from the
package's extdata directory to a specified test directory. After copying, it processes a sample
TSV file and then calls the \code{cortar} function.
}
\examples{
\dontrun{
run_cortar_test("/path/to/your/test/directory")
}
}
