% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0 run_cortar.R
\name{cortar}
\alias{cortar}
\title{Run cortar}
\usage{
cortar(
  file,
  assembly,
  annotation,
  paired,
  stranded,
  subset = FALSE,
  output_dir = "~"
)
}
\arguments{
\item{file}{A file path, pointing to the cortar samplefile.}

\item{assembly}{Assembly used for alignment: either \code{"hg38"} or \code{"hg19"}}

\item{annotation}{Annotation used for alignment: either \code{"UCSC"} or
\code{"1000genomes"}}

\item{paired}{Is the RNA-seq paired-end?: \code{TRUE}/\code{FALSE}}

\item{stranded}{Strandedness of the RNA-seq: \code{0} for unstranded, \code{1}
for forward stranded or \code{2} for reverse stranded}

\item{subset}{Does the RNA-Seq need to be subsetted to the genes
of interest? (Optional, but improves speed of subsequent runs.)
\code{TRUE}/\code{FALSE}}

\item{output_dir}{A directory path, pointing to the desired location for
export of cortar results (e.g. \code{"output/"})}
}
\description{
\code{cortar()} runs the entire cortar pipeline with specified parameters,
and returns excel & pdf reports.
}
\examples{
#### == COMING SOON == ####

}
