% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1 select_gene_and_transcript.R
\name{selectGenesTranscripts}
\alias{selectGenesTranscripts}
\title{Select Genes & Transcripts}
\usage{
selectGenesTranscripts(genes, assembly, annotation)
}
\arguments{
\item{genes}{A character vector of gene names and/or RefSeq transcript IDs
with version numbers (e.g. NM_004006.2)}

\item{assembly}{Assembly used for alignment: either \code{"hg38"} or \code{"hg19"}}

\item{annotation}{Annotation used for alignment: either \code{"UCSC"} or
\code{"1000genomes"}}
}
\value{
A list of GRanges for:
\itemize{
\item genes of interest
\item introns in genes of interest
\item introns in genes of interest (other transcripts)
\item exon-intron junctions of genes of interest
}
}
\description{
\code{selectGenesTranscripts} extracts genes and/or transcripts from the cortar
samplefile and returns gene and intron GRange objects required for further
analysis.
}
\examples{
#### == COMING SOON == ####

}
