% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subsetBamfiles}
\alias{subsetBamfiles}
\title{Extract gene coordinates for RNA-seq subsetting}
\usage{
subsetBamfiles(genes, hg, overhang = 1000)
}
\arguments{
\item{genes}{A character vector with the entrez_gene_symbols,
RefSeq transcript ids, or Ensembl gene or transcript ids for the genes for
analysis.}

\item{hg}{Either 19 or 38}

\item{overhang}{Number of nucleotides flanking the gene to be included
(default: 1000nt)}
}
\value{
A character vector.
}
\description{
Returns the gene coordinates for a given entrez_gene_symbol or
RefSeq/Ensembl transcript to be used for RNA-seq subsetting
}
\examples{
== COMING SOON==

}
